requirements.txt
Keras==2.6.0
numpy==1.19.5
matplotlib==3.4.3
Pillow==8.3.1
opencv-python==4.5.3.56
keras-radam==0.16.0
